refactor: developerNetwork
